refactor(src/components): fixes relative imports, removes unused declarations
