+++ src/components/DateInput.jsx
@@ -1,6 +1,4 @@
-import React from "react";
-import { useGlobalContext } from "../context/AppContext";
-import { Input, Heading, Stack } from "@chakra-ui/react";
+import { Heading, Input, Stack } from "@chakra-ui/react";
 
 // Deconstructing data gotten from the Invoice App and context.jsx. Date on the invoice page
 const DateInput = ({
